Added timeout logic to tests to make them more robust if the cluster is loaded. They should skip the test if conditions are really bad.

* commit 'aacfbf621204075ec96ba3d97b38c68e61065c91':
  Updated the qstat-based timeout tests as per Joe's & Toms comments. Extended its use to test_integration because those tests timed out once as well (but now they don't)
  Extended new timeout wrapper to test_sge_jobs
  Partial work
  Fixed bug in test_sge_test whereby it tried to write to Tom's home directory :-) Added progressive timeouts - if the cluster is happy the test will be faster. If it is slow, it wauits longer and eventually prints a qstat.

+++ tests/test_sge_jobs.py
@@ -1,5 +1,11 @@
 import pytest
-from time import sleep
+import sys
+
+if sys.version_info < (3, 0):
+    from functools32 import partial
+else:
+    from functools import partial
+
 from datetime import datetime, timedelta
 
 from jobmon import sge
@@ -7,6 +13,8 @@ from jobmon.database import session_scope
 from jobmon.models import JobInstance, JobInstanceStatus
 from jobmon.job_instance_factory import execute_sge
 from jobmon.job_list_manager import JobListManager
+
+from tests.timeout_and_skip import timeout_and_skip
 
 
 @pytest.fixture(scope='function')
@@ -23,12 +31,19 @@ def test_valid_command(dag_id, job_list_manager_sge):
         "sge_foobar", slots=2, mem_free=4, project="proj_qlogins",
         max_attempts=3)
     job_list_manager_sge.queue_job(job_id)
-    sleep(60)  # Give some time for the job to get to the executor
+
+    timeout_and_skip(10, 120, 1, partial(
+        valid_command_check,
+        job_list_manager_sge=job_list_manager_sge))
+
+
+def valid_command_check(job_list_manager_sge):
     done = job_list_manager_sge.get_new_done()
-    fail_msg = ("Test failed: check that you have permission to run under "
-                "'proj_qlogins' and that there are available jobs under this"
-                " project")
-    assert len(done) == 1, fail_msg
+    if len(done) == 1:
+        # Success
+        return True
+    else:
+        return False
 
 
 def test_context_args(jsm_jqs, job_list_manager_sge):
@@ -38,17 +53,27 @@ def test_context_args(jsm_jqs, job_list_manager_sge):
         "sge_foobar", slots=2, mem_free=4, max_attempts=3,
         context_args={'sge_add_args': '-a {}'.format(delay_to)})
     job_list_manager_sge.queue_job(job_id)
-    sleep(60)  # Give some time for the job to get instantiated
+
+    timeout_and_skip(10, 180, 1, partial(
+        context_args_check,
+        job_id=job_id))
+
+
+def context_args_check(job_id):
     with session_scope() as session:
         jis = session.query(JobInstance).filter_by(job_id=job_id).all()
         njis = len(jis)
         status = jis[0].status
         sge_jid = jis[0].executor_id
-
     # Make sure the job actually got to SGE
-    assert njis == 1
-    # Make sure it hasn't advanced to running (i.e. the -a argument worked)
-    assert status == JobInstanceStatus.SUBMITTED_TO_BATCH_EXECUTOR
+    if njis == 1:
+        # Make sure it hasn't advanced to running (i.e. the -a argument worked)
+        assert status == JobInstanceStatus.SUBMITTED_TO_BATCH_EXECUTOR
+        # Cleanup
+        sge.qdel(sge_jid)
+        return True
+    else:
+        return False
 
-    # Cleanup
-    sge.qdel(sge_jid)
+
+
